System Overhaul
Rewrote the whole system in order to act more OOP-Friendly. Trying to
keep it fresh and to follow a better code standard.
